+++ compiler/src/test/java/info/teksol/mc/mindcode/compiler/ast/AstIndentedPrinterTest.java
@@ -149,7 +149,8 @@ public class AstIndentedPrinterTest extends AbstractAstBuilderTest {
                                                         refModifier = false
                                                     }
                                                 }
-                                            }
+                                            },
+                                            elseDefined = true
                                         }
                                     },
                                     callType = NONE